validate in service

+++ services/src/main/java/ar/edu/itba/paw/services/GameServiceImpl.java
@@ -1,17 +1,22 @@
 package ar.edu.itba.paw.services;
 
+import ar.edu.itba.paw.exceptions.GameAlreadyExist;
 import ar.edu.itba.paw.exceptions.GameHasNotBeenPlayException;
 import ar.edu.itba.paw.exceptions.GameNotFoundException;
+import ar.edu.itba.paw.exceptions.InvalidGameKeyException;
 import ar.edu.itba.paw.exceptions.TeamFullException;
+import ar.edu.itba.paw.exceptions.UnauthorizedExecption;
 import ar.edu.itba.paw.interfaces.GameDao;
 import ar.edu.itba.paw.interfaces.GameService;
 import ar.edu.itba.paw.interfaces.PremiumUserService;
+import ar.edu.itba.paw.interfaces.SessionService;
 import ar.edu.itba.paw.interfaces.TeamService;
 import ar.edu.itba.paw.interfaces.UserService;
 import ar.edu.itba.paw.models.Game;
 import ar.edu.itba.paw.models.GameKey;
 import ar.edu.itba.paw.models.GameSort;
 import ar.edu.itba.paw.models.Page;
+import ar.edu.itba.paw.models.PremiumUser;
 import ar.edu.itba.paw.models.Team;
 import ar.edu.itba.paw.models.User;
 import org.slf4j.Logger;
@@ -22,7 +27,6 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class GameServiceImpl implements GameService {
@@ -41,6 +45,9 @@ public class GameServiceImpl implements GameService {
     @Autowired
     private PremiumUserService premiumUserService;
 
+    @Autowired
+    private SessionService sessionService;
+
     private static final String GROUP       = "Group";
     private static final String INDIVIDUAL  = "Individual";
     private static final String COMPETITIVE = "Competitive";
@@ -51,20 +58,26 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
-    public Optional<Game> create(final String teamName1, final String teamName2, final LocalDateTime startTime,
+    public Game create(final String teamName1, final String teamName2, final LocalDateTime startTime,
                                  final long durationInMinutes, final boolean isCompetitive, final boolean isIndividual,
                                  final String country, final String state, final String city,
                                  final String street, final String tornamentName, final String description,
                                  final String title) {
+        //TODO: validate that teams exist and maybe that logged user has permissions
         String type = (isIndividual ? INDIVIDUAL : GROUP) + '-' + (isCompetitive ? COMPETITIVE : FRIENDLY);
 
         return gameDao.create(teamName1, teamName2, startTime.toString(),
                 startTime.plusMinutes(durationInMinutes).toString(), type, null,
-                country, state, city, street, tornamentName, description, title);
+                country, state, city, street, tornamentName, description, title)
+                .orElseThrow(() -> {
+                    GameKey gameKey = new GameKey(startTime, teamName1, startTime.plusMinutes(durationInMinutes));
+                    LOGGER.trace("Creation fails, match '{}' already exist", gameKey.toString());
+                    return new GameAlreadyExist("Creation fails, match '" + gameKey.toString() + "' already exist");
+                });
     }
 
     @Override
-    public Optional<Game> createNoTeamGame(final LocalDateTime startTime, final long durationInMinutes,
+    public Game createNoTeamGame(final LocalDateTime startTime, final long durationInMinutes,
                                            final boolean isCompetitive, final String country,
                                            final String state, final String city,
                                            final String street, final String tornamentName,
@@ -73,26 +86,33 @@ public class GameServiceImpl implements GameService {
         Team team1 = teamService.createTempTeam1(creatorName, creatorId, sportName);
         Team team2 = teamService.createTempTeam2(creatorName, creatorId, sportName);
 
-        Optional<Game> game = create(team1.getName(), team2.getName(), startTime, durationInMinutes, isCompetitive,
-                true, country, state, city, street, tornamentName, description, title);
-        return insertUserInGameTeam(game.orElse(null), creatorId, true);
+        String type = INDIVIDUAL + '-' + (isCompetitive ? COMPETITIVE : FRIENDLY);
+
+        Game game = gameDao.create(team1.getName(), team2.getName(), startTime.toString(),
+                startTime.plusMinutes(durationInMinutes).toString(), type, null,
+                country, state, city, street, tornamentName, description, title).orElseThrow(() -> {
+                    GameKey gameKey = new GameKey(startTime, team1.getName(), startTime.plusMinutes(durationInMinutes));
+                    LOGGER.trace("Creation fails, match '{}' already exist", gameKey.toString());
+                    return new GameAlreadyExist("Creation fails, match '" + gameKey.toString() + "' already exist");
+                });
+        return insertUserInGameTeam(game, creatorId, true);
     }
 
     @Override
-    public Optional<Game> insertUserInGame(final String key, final long userId) {
-        Optional<Game> game;
+    public Game insertUserInGame(final String key, final long userId) {
+        Game game;
         try {
-            game = insertUserInGameTeam(findByKey(key).orElse(null), userId, true);
+            game = insertUserInGameTeam(findByKey(key), userId, true);
         }
         catch (TeamFullException e) {
-            game = insertUserInGameTeam(findByKey(key).orElse(null), userId, false);
+            game = insertUserInGameTeam(findByKey(key), userId, false);
         }
         return game;
     }
 
     @Override
     public boolean deleteUserInGame(final String key, final long userId) {
-        Game game = findByKey(key).orElseThrow(() -> new GameNotFoundException("Game '" + key + "' was not found"));
+        Game game = findByKey(key);
         for (User user : game.getTeam1().getPlayers()) {
             if (user.getUserId() == userId) {
                 LOGGER.trace("Found user: {} in team1", userId);
@@ -113,7 +133,7 @@ public class GameServiceImpl implements GameService {
                 return true;
             }
         }
-        LOGGER.trace("Not found user: {} in game", userId);
+        LOGGER.trace("Not found user: '{}' in match '{}'", userId, key);
         return false;
     }
 
@@ -146,47 +166,57 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
-    public Optional<Game> modify(final String teamName1, final String teamName2, final String startTime,
+    public Game modify(final String teamName1, final String teamName2, final String startTime,
                                  final Long minutesOfDuration, final String type, final String result,
                                  final String country, final String state, final String city,
                                  final String street, final String tornamentName, final String description,
                                  final String title, final String key) {
-        GameKey gameKey = new GameKey(key);
-        Optional<Game> gameOptional = findByKey(key), game = Optional.empty();
-        if (gameOptional.isPresent()) {
-            if ((teamName1 != null || teamName2 != null) && gameOptional.get().getGroupType().equals(INDIVIDUAL)) {
-                throw new IllegalArgumentException("Cannot modify teams in a individual match");
-            }
-            game = gameDao.modify(teamName1, teamName2, startTime, (minutesOfDuration != null) ?
-                            gameOptional.get().getStartTime().plusMinutes(minutesOfDuration).toString() : null, type,
-                    result, country, state, city, street, tornamentName, description, title, gameKey.getTeamName1(),
-                    gameKey.getStartTime(), gameKey.getFinishTime());
+        GameKey gameKey = getGameKey(key);
+        Game gameOptional = findByKey(key), game;
+        if ((teamName1 != null || teamName2 != null) && gameOptional.getGroupType().equals(INDIVIDUAL)) {
+            throw new IllegalArgumentException("Cannot modify teams in a individual match");
         }
+        game = gameDao.modify(teamName1, teamName2, startTime, (minutesOfDuration != null) ?
+                        gameOptional.getStartTime().plusMinutes(minutesOfDuration).toString() : null, type,
+                result, country, state, city, street, tornamentName, description, title, gameKey.getTeamName1(),
+                gameKey.getStartTime(), gameKey.getFinishTime()).orElseThrow(() -> {
+                    LOGGER.trace("Modify fails, match '{}' not found", key);
+                    return new GameNotFoundException("Modify fails, match '" + key + "' not found");
+                });
+
         return game;
     }
 
     @Override
     public boolean remove(final String key) {
-        GameKey gameKey = new GameKey(key);
+        GameKey gameKey = getGameKey(key);
+        PremiumUser premiumUser = sessionService.getLoggedUser().get();//TODO maybe check
+        if (!findByKey(key).getPrimaryKey().getTeam1().getLeader().equals(premiumUser)) {
+            LOGGER.trace("User '{}' is not creator of '{}' match", premiumUser.getUserName(), key);
+            throw new UnauthorizedExecption("User '" + premiumUser.getUserName() +
+                    "' is not creator of '" + key + "' match");//TODO maybe another exception
+        }
         return gameDao.remove(gameKey.getTeamName1(), gameKey.getStartTime(), gameKey.getFinishTime());
     }
 
     @Override
-    public Optional<Game> findByKey(final String key) {
-        GameKey gameKey = new GameKey(key);
-        return gameDao.findByKey(gameKey.getTeamName1(), gameKey.getStartTime(), gameKey.getFinishTime());
+    public Game findByKey(final String key) {
+        GameKey gameKey = getGameKey(key);
+        return gameDao.findByKey(gameKey.getTeamName1(), gameKey.getStartTime(), gameKey.getFinishTime())
+                .orElseThrow(() -> {
+                    LOGGER.trace("Find by key fails, match '{}' not found", key);
+                    return new GameNotFoundException("Find by key fails, match '" + key + "' not found");
+                });
     }
 
     @Override
-    public Optional<Game> updateResultOfGame(final String key, final int scoreTeam1, final int scoreTeam2) {
-        GameKey gameKey = new GameKey(key);
-        Optional<Game> game = gameDao.findByKey(gameKey.getTeamName1(), gameKey.getStartTime(), gameKey.getFinishTime());
-        if (game.isPresent()) {
-            if (game.get().getFinishTime().compareTo(LocalDateTime.now()) > 0) {
-                throw new GameHasNotBeenPlayException("The game has not been play");
-            }
-            game.get().setResult(scoreTeam1 + "-" + scoreTeam2);
+    public Game updateResultOfGame(final String key, final int scoreTeam1, final int scoreTeam2) {
+       Game game = findByKey(key);
+        if (game.getFinishTime().compareTo(LocalDateTime.now()) > 0) {
+            throw new GameHasNotBeenPlayException("The game has not been play");
         }
+        game.setResult(scoreTeam1 + "-" + scoreTeam2);
+
         return game;
     }
 
@@ -198,14 +228,21 @@ public class GameServiceImpl implements GameService {
         return listsOfGames;
     }
 
-    private Optional<Game> insertUserInGameTeam(final Game game, final long userId, final boolean toTeam1) {
-        if (game != null) {
-            if (!toTeam1) {
-                game.setTeam2(teamService.addPlayer(game.team2Name(), userId));
-            } else {
-                game.getPrimaryKey().setTeam1(teamService.addPlayer(game.team1Name(), userId));
-            }
+    private Game insertUserInGameTeam(final Game game, final long userId, final boolean toTeam1) {
+        if (!toTeam1) {
+            game.setTeam2(teamService.addPlayer(game.team2Name(), userId));
+        } else {
+            game.getPrimaryKey().setTeam1(teamService.addPlayer(game.team1Name(), userId));
         }
-        return Optional.of(game);//TODO: check
+        return game;
+    }
+
+    private GameKey getGameKey(final String key) {
+        try {
+            return new GameKey(key);
+        }
+        catch (Exception e) {
+            throw new InvalidGameKeyException("Invalid key");
+        }
     }
 }
